wip: update after package upgrade

+++ src/renderer/components/main-space/workspace/enrichment/enrichment-container.tsx
@@ -4,26 +4,19 @@ import React, { useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import { commitAction } from "../../../../reducers/enhancers/undoable/undoable-actions";
-import {
-  markAsToDelete,
-  unmarkAsToDelete,
-} from "../../../../reducers/files-and-folders/files-and-folders-actions";
+import { markAsToDelete, unmarkAsToDelete } from "../../../../reducers/files-and-folders/files-and-folders-actions";
 import {
   getCommentsFromStore,
   getElementsToDeleteFromStore,
   getFilesAndFoldersFromStore,
 } from "../../../../reducers/files-and-folders/files-and-folders-selectors";
 import { updateCommentThunk } from "../../../../reducers/files-and-folders/files-and-folders-thunks";
-import type { FilesAndFoldersMap } from "../../../../reducers/files-and-folders/files-and-folders-types";
+import { type FilesAndFoldersMap } from "../../../../reducers/files-and-folders/files-and-folders-types";
 import { getFilesAndFoldersMetadataFromStore } from "../../../../reducers/files-and-folders-metadata/files-and-folders-metadata-selectors";
-import type { FilesAndFoldersMetadataMap } from "../../../../reducers/files-and-folders-metadata/files-and-folders-metadata-types";
-import type { StoreState } from "../../../../reducers/store";
+import { type FilesAndFoldersMetadataMap } from "../../../../reducers/files-and-folders-metadata/files-and-folders-metadata-types";
+import { type StoreState } from "../../../../reducers/store";
 import { addTag, untagFile } from "../../../../reducers/tags/tags-actions";
-import {
-  getAllTagIdsForFile,
-  getTagsByIds,
-  getTagsFromStore,
-} from "../../../../reducers/tags/tags-selectors";
+import { getAllTagIdsForFile, getTagsByIds, getTagsFromStore } from "../../../../reducers/tags/tags-selectors";
 import { useWorkspaceMetadata } from "../../../../reducers/workspace-metadata/workspace-metadata-selectors";
 import { isFolder } from "../../../../utils";
 import { getAllChildren } from "../../../../utils/file-and-folders";
@@ -32,7 +25,7 @@ import { Enrichment } from "./enrichment";
 const computeTreeSize = (
   filesAndFoldersMetadataMap: FilesAndFoldersMetadataMap,
   filesAndFoldersMap: FilesAndFoldersMap,
-  filesAndFoldersId: string
+  filesAndFoldersId: string,
 ) => {
   const filesAndFolders = filesAndFoldersMap[filesAndFoldersId];
   return isFolder(filesAndFolders)
@@ -44,18 +37,11 @@ const handleTracking = (
   isCurrentFileMarkedToDelete: boolean,
   filesAndFoldersMetadataMap: FilesAndFoldersMetadataMap,
   filesAndFoldersMap: FilesAndFoldersMap,
-  filesAndFoldersId: string
+  filesAndFoldersId: string,
 ): void => {
   if (!isCurrentFileMarkedToDelete) {
-    const sizeRaw = computeTreeSize(
-      filesAndFoldersMetadataMap,
-      filesAndFoldersMap,
-      filesAndFoldersId
-    );
-    const fileCount = getAllChildren(
-      filesAndFoldersMap,
-      filesAndFoldersId
-    ).length;
+    const sizeRaw = computeTreeSize(filesAndFoldersMetadataMap, filesAndFoldersMap, filesAndFoldersId);
+    const fileCount = getAllChildren(filesAndFoldersMap, filesAndFoldersId).length;
 
     getTrackerProvider().track("Feat(4.0) Element Marked To Delete", {
       fileCount,
@@ -73,16 +59,14 @@ export const EnrichmentContainer: React.FC = () => {
   const filesAndFoldersId = lockedElementId || hoveredElementId;
 
   const filesAndFoldersMap = useSelector(getFilesAndFoldersFromStore);
-  const filesAndFoldersMetadataMap = useSelector(
-    getFilesAndFoldersMetadataFromStore
-  );
+  const filesAndFoldersMetadataMap = useSelector(getFilesAndFoldersMetadataFromStore);
 
   const createTag = useCallback(
     (tagName: string, ffId: string) => {
       dispatch(addTag(tagName, ffId));
       dispatch(commitAction());
     },
-    [dispatch]
+    [dispatch],
   );
 
   const untag = useCallback(
@@ -90,7 +74,7 @@ export const EnrichmentContainer: React.FC = () => {
       dispatch(untagFile(tagName, ffId));
       dispatch(commitAction());
     },
-    [dispatch]
+    [dispatch],
   );
 
   const updateComment = useCallback(
@@ -98,40 +82,27 @@ export const EnrichmentContainer: React.FC = () => {
       dispatch(updateCommentThunk(filesAndFoldersId, comments));
       dispatch(commitAction());
     },
-    [dispatch, filesAndFoldersId]
+    [dispatch, filesAndFoldersId],
   );
 
-  const currentFileComment =
-    useSelector(getCommentsFromStore)[filesAndFoldersId] || "";
+  const currentFileComment = useSelector(getCommentsFromStore)[filesAndFoldersId] || "";
 
   const tagIdsForCurrentFile = useSelector((state: StoreState) =>
-    getAllTagIdsForFile(getTagsFromStore(state), filesAndFoldersId)
+    getAllTagIdsForFile(getTagsFromStore(state), filesAndFoldersId),
   );
 
   const tagsForCurrentFile = useSelector((state: StoreState) =>
-    getTagsByIds(getTagsFromStore(state), tagIdsForCurrentFile)
+    getTagsByIds(getTagsFromStore(state), tagIdsForCurrentFile),
   );
 
   const filesToDelete = useSelector(getElementsToDeleteFromStore);
   const isCurrentFileMarkedToDelete = filesToDelete.includes(filesAndFoldersId);
 
   const toggleCurrentFileDeleteState = useCallback(() => {
-    handleTracking(
-      isCurrentFileMarkedToDelete,
-      filesAndFoldersMetadataMap,
-      filesAndFoldersMap,
-      filesAndFoldersId
-    );
-    if (isCurrentFileMarkedToDelete)
-      dispatch(unmarkAsToDelete(filesAndFoldersId));
+    handleTracking(isCurrentFileMarkedToDelete, filesAndFoldersMetadataMap, filesAndFoldersMap, filesAndFoldersId);
+    if (isCurrentFileMarkedToDelete) dispatch(unmarkAsToDelete(filesAndFoldersId));
     else dispatch(markAsToDelete(filesAndFoldersId));
-  }, [
-    dispatch,
-    isCurrentFileMarkedToDelete,
-    filesAndFoldersId,
-    filesAndFoldersMap,
-    filesAndFoldersMetadataMap,
-  ]);
+  }, [dispatch, isCurrentFileMarkedToDelete, filesAndFoldersId, filesAndFoldersMap, filesAndFoldersMetadataMap]);
 
   const isFocused = filesAndFoldersId !== "";
   const isLocked = lockedElementId !== "";
